backend_fastapi/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
pymysql>=1.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pydantic>=2.6
pydantic-settings>=2.2
